New Update
-Move all stuff related to save/load, data and storage to uml_storage_manager.py
-Move all methods in interface to uml_manager.py
-Now the code organization is clearer
- I also made some small changes so that it will prompt message when user trying to perform any commands that print things on the terminal when the program is blank
