+++ UML_CORE/UML_CLASS/uml_class.py
@@ -14,19 +14,19 @@ List of date modified:
 ################################################################
 # IMPORTED MODULES #
 
-import UML_MANAGER.uml_manager as UML_MANAGER
+import UML_MANAGER.uml_storage_manager as UML_STORAGE_MANAGER
 from UML_UTILITY.FORMAT_CHECKING.validators import check_format
 
 ################################################################
 
 # GET DATA FROM JSON FILE #
-data_list = UML_MANAGER.data_list
+data_list = UML_STORAGE_MANAGER.data_list
 # GET CLASS AND ITS ATTRIBUTES LIST #
-class_and_attr_list = UML_MANAGER.class_and_attr_list
+class_and_attr_list = UML_STORAGE_MANAGER.class_and_attr_list
 # GET RELATIONSHIP LIST #
-relationship_list = UML_MANAGER.relationship_list
+relationship_list = UML_STORAGE_MANAGER.relationship_list
 # GET CLASS NAME LIST #
-class_list = UML_MANAGER.class_list
+class_list = UML_STORAGE_MANAGER.class_list
 
 ################################################################################
 # WORKING WITH CLASSES #
