+++ UML_UTILITY/SAVED_FILES/NAME_LIST.json
@@ -10,5 +10,8 @@
     },
     {
         "data_clone": "off"
+    },
+    {
+        "test02": "off"
     }
 ]
